Extract test classes into their own files
